chore: insert asset encoding
